fix: added unit tests for debt swap
